# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puryaideu V2 API",
        "description": "Puryaideu APIs documentation",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/review/create": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Create Review",
                "operationId": "c08c6afa37e70f5f2770f647f2eaed69",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "booking_id": 1,
                                    "rate": 3,
                                    "comment": "Ride was smooth!",
                                    "reviewed_by_role": "customer"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Ride Review Successful!",
                                        "review": {
                                            "booking_id": 1,
                                            "rate": 3,
                                            "comment": "Ride was smooth!",
                                            "reviewed_by_role": "customer",
                                            "ride_date": "2021-11-17T07:42:19.000000Z",
                                            "user_id": 3,
                                            "rider_id": 1,
                                            "updated_at": "2021-11-17T08:21:19.000000Z",
                                            "created_at": "2021-11-17T08:21:19.000000Z",
                                            "id": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden Access!"
                    },
                    "500": {
                        "description": "Internal Server Error!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://puryaideu_v2.test/oauth/authorize",
                        "tokenUrl": "http://puryaideu_v2.test/oauth/token",
                        "refreshUrl": "http://puryaideu_v2.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}